requirements.txt
llama-index == 0.8.56
openai == 0.27.10
pydantic == 1.10.9
fastapi == 0.97.0
uvicron == 3.10.11
python-dotenv==1.0.0
